feat: enhance QuickAccess and Screen components; add badge functionality and improve layout styles
